refactor(web): mini-module resource for timedPromise
